TD-5216 Moves to import namespace and separates form data in viewmodel
