+++ DigitalLearningSolutions.Web/ViewModels/Frameworks/Import/ImportCompetenciesFormData.cs
@@ -1,13 +1,11 @@
-﻿namespace DigitalLearningSolutions.Web.ViewModels.Frameworks
+﻿namespace DigitalLearningSolutions.Web.ViewModels.Frameworks.Import
 {
     using DigitalLearningSolutions.Web.Attributes;
     using Microsoft.AspNetCore.Http;
     using System.ComponentModel.DataAnnotations;
 
-    public class ImportCompetenciesViewModel
+    public class ImportCompetenciesFormData
     {
-        public int FrameworkId { get; set; }
-        public bool IsNotBlank { get; set; }
         [Required(ErrorMessage = "Import competencies file is required")]
         [AllowedExtensions(new[] { ".xlsx" }, "Import competencies file must be in xlsx format")]
         [MaxFileSize(5 * 1024 * 1024, "Maximum allowed file size is 5MB")]
