+++ DigitalLearningSolutions.Web/Views/Frameworks/Developer/Import/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@using DigitalLearningSolutions.Web.Extensions
-@using DigitalLearningSolutions.Web.ViewModels.Frameworks
+@using DigitalLearningSolutions.Web.ViewModels.Frameworks.Import
 @model ImportCompetenciesViewModel
 @{
     ViewData["Application"] = "Framework Service";
@@ -31,6 +31,10 @@
             <vc:error-summary order-of-property-names="@(new[] { nameof(Model.ImportFile) })" />
         }
         <h1 class="nhsuk-heading-xl">Bulk upload @(Model.IsNotBlank ? "or update" : "") competencies</h1>
+        @if (Model.Framework.PublishStatusID == 3)
+        {
+            <partial name="Shared/_PublishedWarning" />
+        }
         <vc:inset-text css-class="" text="The building blocks for your framework will be referred to as 'competencies' for the purpose of importing. Once they have been imported, they will be referred to using your chosen vocabulary."></vc:inset-text>
         <div class="nhsuk-form-group">
             @if (Model.IsNotBlank)
@@ -101,7 +105,7 @@
             Once you have an Excel competencies workbook, add or update competencies to the worksheet, save and start the upload process.
         </p>
         <form class="nhsuk-u-margin-bottom-3" method="post" enctype="multipart/form-data">
-            <input type="hidden" asp-for="@Model.FrameworkId" />
+            <input type="hidden" asp-for="@Model.Framework.ID" />
             <vc:file-input asp-for="@nameof(Model.ImportFile)" label="Excel file ready for import" hint-text="" css-class="nhsuk-u-width-one-half" />
             <button class="nhsuk-button" type="submit">Start upload</button>
         </form>
