+++ DigitalLearningSolutions.Web/Views/Frameworks/Developer/Import/UploadResults.cshtml
@@ -1,6 +1,6 @@
 ﻿@inject IConfiguration Configuration
 @using DigitalLearningSolutions.Web.Extensions
-@using DigitalLearningSolutions.Web.ViewModels.Frameworks
+@using DigitalLearningSolutions.Web.ViewModels.Frameworks.Import
 @using Microsoft.Extensions.Configuration
 @model ImportCompetenciesResultsViewModel
 @{
